rename: auth 폴더 대소문자 변경
