--- src/pages/AlbumPage.tsx
@@ -1,6 +1,0 @@
-function Album() {
-  return (
-    <div>Album</div>
-  )
-}
-export default Album
